requirements.txt
pandas==2.2.2
pillow==10.3.0
python-dotenv==1.0.1
PyYAML==6.0.1
tqdm==4.66.4
wandb==0.17.3